Basic examples done
I now have three programs baseline, std-baseline and pipeline that do roughly
the same things except for differences that I will eventually document.

+++ src/core.rs
@@ -10,7 +10,7 @@ pub trait Task {
 #[derive(Debug)]
 pub struct Event {
     pub line: Box<str>,
-    pub spaces: Option<usize>,
+    pub spaces: Option<u32>,
 }
 
 pub enum Error {
